requirements.txt
django-ckeditor
xmltodict
qrcode